Ajustes pós review
