+++ src/main/java/br/com/deveficiente/bolaoapi/services/poll/api/InvitationEmailController.java
@@ -6,8 +6,6 @@ import br.com.deveficiente.bolaoapi.services.poll.api.model.InvitationAccepted;
 import br.com.deveficiente.bolaoapi.shared.api.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.mail.MailMessage;
-import org.springframework.mail.SimpleMailMessage;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -16,16 +14,14 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.Optional;
-import java.util.Set;
 
 @RestController
 @RequestMapping("/invitations")
-public class InvitationController {
+public class InvitationEmailController {
 
     private final InvitationRepository invitationRepository;
-//    private JavaMailSender mailSender;
 
-    public InvitationController(InvitationRepository invitationRepository) {
+    public InvitationEmailController(InvitationRepository invitationRepository) {
         this.invitationRepository = invitationRepository;
     }
 
@@ -39,22 +35,4 @@ public class InvitationController {
 
         return ResponseEntity.ok(new InvitationAccepted(invitation.get()));
     }
-
-    public void sendInvitationsByEmail(Set<Invitation> invitations) {
-        invitations.forEach(invitation -> {
-            String ownerLogin = invitation.getPoll().getOwner().getLogin();
-            System.out.println(invitation.getInvitationLink());
-//            mailSender.send(buildMessage(ownerLogin, invitation));
-        });
-    }
-
-    private MailMessage buildMessage(String from, Invitation invitation) {
-        SimpleMailMessage message = new SimpleMailMessage();
-        message.setTo(invitation.getEmail());
-        message.setFrom(from);
-        message.setText(invitation.getInvitationLink());
-
-        return message;
-    }
-
 }
